# services/telemetry-ingestion/api/openapi.yaml
openapi: 3.0.3
info:
  title: AtlasMesh Telemetry Ingestion API
  version: 1.0.0
  description: |
    High-throughput telemetry ingestion service for AtlasMesh Fleet OS.
    
    Handles real-time vehicle telemetry data with:
    - Schema validation against Avro schemas
    - Data quality checks and enrichment
    - Multi-path storage (hot/cold) routing
    - Dead letter queue handling for failed messages
    - Real-time streaming via WebSocket
    
    ## Performance Characteristics
    - Ingestion rate: 100,000+ messages/second
    - P99 latency: <100ms for batch ingestion
    - P99 latency: <10ms for streaming ingestion
    - Availability: 99.95%
    
    ## Data Formats
    - JSON: For development and testing
    - Avro: For production (recommended for performance)
    - Protobuf: For edge devices with bandwidth constraints
    
    ## Rate Limits
    - Authenticated vehicles: 10,000 requests/minute
    - Batch ingestion: 1,000 messages per request
    - Streaming: Unlimited for authenticated vehicles

  contact:
    name: AtlasMesh Telemetry Team
    email: telemetry-team@atlasmesh.ai
    url: https://docs.atlasmesh.ai/telemetry
  license:
    name: Proprietary
    url: https://atlasmesh.ai/license
servers:
  - url: http://telemetry-ingestion:8095/v1
    description: Internal service endpoint
  - url: https://telemetry.atlasmesh.ai/v1
    description: Production telemetry ingestion
...
    description: Staging telemetry ingestion
security:
  - vehicleAuth: []
  - serviceAuth: []
paths:
...
  /schemas:
    get:
      summary: List available telemetry schemas
      operationId: listSchemas
      tags: [Schema Management]
      parameters:
        - name: version
          in: query
          schema:
            type: string
          description: Filter by schema version
        - name: format
          in: query
          schema:
            type: string
            enum: [avro, json-schema, protobuf]
          description: Schema format
      responses:
        '200':
          description: List of available schemas
          content:
            application/json:
              schema:
                type: object
                properties:
                  schemas:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchemaInfo'
  /schemas/{schemaId}:
    get:
      summary: Get specific schema definition
      operationId: getSchema
      tags: [Schema Management]
      parameters:
        - name: schemaId
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [avro, json-schema, protobuf]
            default: avro
      responses:
        '200':
          description: Schema definition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaDefinition'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    SchemaInfo:
      type: object
      properties:
        schema_id:
          type: string
        name:
          type: string
        version:
          type: string
        format:
          type: string
          enum: [avro, json-schema, protobuf]
        created_at:
          type: string
          format: date-time
        description:
          type: string
    SchemaDefinition:
      type: object
      properties:
        schema_id:
          type: string
        name:
          type: string
        version:
          type: string
        format:
          type: string
        definition:
          description: Schema definition (format depends on schema type)
        examples:
          type: array
          items:
            type: object
          description: Example data conforming to schema
...
    ErrorResponse:
      type: object
      required: [error, message, timestamp, correlation_id]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
        timestamp:
          type: string
          format: date-time
        correlation_id:
          type: string
          description: Request correlation ID for tracing
        retry_after:
          type: integer
          description: Seconds to wait before retrying
        documentation_url:
          type: string
          format: uri
          description: Link to relevant documentation
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    vehicleAuth:
      type: http
      scheme: bearer
      bearerFormat: Vehicle-Certificate
      description: Vehicle authentication using X.509 certificates
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: Service-Token
      description: Service-to-service authentication